requirements.txt
python-dotenv
discord.py
openai
yt_dlp
dacite
PyYAML
PyNaCl
async_timeout
dpytest
pytest